+++ examples/subscriptions/package.json
@@ -8,6 +8,7 @@
     },
     "dependencies": {
         "@modelcontextprotocol/client": "workspace:*",
+        "@modelcontextprotocol/node": "workspace:*",
         "@modelcontextprotocol/server": "workspace:*",
         "zod": "catalog:runtimeShared"
     },
