+++ packages/middleware/node/src/index.ts
@@ -1,3 +1,11 @@
 export * from './middleware/hostHeaderValidation.js';
 export * from './middleware/originValidation.js';
 export * from './streamableHttp.js';
+export type {
+    FetchLikeMcpHandler,
+    NodeIncomingMessageLike,
+    NodeMcpRequestHandler,
+    NodeServerResponseLike,
+    ToNodeHandlerOptions
+} from './toNodeHandler.js';
+export { toNodeHandler } from './toNodeHandler.js';
